feat(Locust): Mejoras en la generación de carga y documentación de funciones
chore(Locust): Se vacía el archivo `usuarios_registrados.json` al iniciar la prueba.  Se añaden docstrings a las funciones de `generar_carga.py`.

refactor(web): Se refactoriza la obtención del ID de carrito de usuario en `carritoController.py` y `pedidoController.py`. Se agrega la eliminación de items del carrito luego del checkout en `pedidoController.py`. Se modifica el flujo de `scripts.js` luego de un pedido exitoso. Se refactoriza la obtención del ID de carrito en `carrito.py`.
