+++ Locust/generar_carga.py
@@ -1,12 +1,21 @@
 from locust import HttpUser, task, between
 import random, json, string, os
 
+# Vaciar archivo JSON de usuarios si existe al iniciar
+if os.path.exists("usuarios_registrados.json"):
+    with open("usuarios_registrados.json", "w") as f:
+        f.write("")
+
 # === Archivos de persistencia ===
 ARCHIVO_USUARIOS = "usuarios_registrados.json"
 ARCHIVO_PRODUCTOS = "productos_cache.json"
 
 # === Funciones auxiliares ===
 def generar_usuario():
+    """Genera un usuario con datos aleatorios
+
+    Return: dict con datos del usuario
+    """
     nombre = ''.join(random.choices(string.ascii_lowercase, k=10))
     return {
         "nombre": nombre,
@@ -17,16 +26,35 @@ def generar_usuario():
     }
 
 def cargar_json_lines(path):
+    """Carga un archivo JSON línea a línea
+
+    Keyword arguments:
+    path -- ruta del archivo
+    Return: lista de objetos JSON cargados
+    """
     if os.path.exists(path):
         with open(path, "r") as f:
             return [json.loads(line.strip()) for line in f if line.strip()]
     return []
 
 def guardar_json_line(path, data):
+    """Guarda un diccionario como JSON en una nueva línea en el archivo
+
+    Keyword arguments:
+    path -- ruta del archivo
+    data -- diccionario a guardar
+    Return: None
+    """
     with open(path, "a") as f:
         f.write(json.dumps(data) + "\n")
 
 def cargar_json(path):
+    """Carga un archivo JSON completo
+
+    Keyword arguments:
+    path -- ruta del archivo
+    Return: objeto JSON o lista vacía
+    """
     if os.path.exists(path):
         with open(path, "r") as f:
             return json.load(f)
@@ -37,6 +65,10 @@ class MiUsuario(HttpUser):
     wait_time = between(1, 2)
 
     def on_start(self):
+        """Se ejecuta al iniciar cada usuario de prueba
+
+        Return: None
+        """
         self.usuario = generar_usuario()
         self.usuarios_registrados = cargar_json_lines(ARCHIVO_USUARIOS)
         self.autenticado = False
@@ -44,6 +76,13 @@ class MiUsuario(HttpUser):
         self.lista_pedidos = []
 
     def manejar_redireccion(self, response, mensaje_ok):
+        """Gestiona respuestas HTTP, chequeando redirecciones y errores
+
+        Keyword arguments:
+        response -- respuesta HTTP
+        mensaje_ok -- mensaje a mostrar si es exitosa
+        Return: True si OK o redirección válida, False si error
+        """
         if response.status_code == 200:
             print(mensaje_ok)
             return True
@@ -59,6 +98,10 @@ class MiUsuario(HttpUser):
 
     @task(2)
     def registro_post(self):
+        """Tarea para registrar un usuario vía POST
+
+        Return: None
+        """
         with self.client.post("/api/registro", data=self.usuario, catch_response=True) as response:
             if response.status_code == 200 and "redirect_url" in response.json():
                 print("✅ Usuario registrado exitosamente")
@@ -70,6 +113,10 @@ class MiUsuario(HttpUser):
 
     @task(2)
     def login_post(self):
+        """Tarea para login con usuario registrado aleatorio
+
+        Return: None
+        """
         if not self.usuarios_registrados:
             return
 
@@ -88,18 +135,34 @@ class MiUsuario(HttpUser):
 
     @task(1)
     def register_get(self):
+        """Tarea para solicitar la página de registro vía GET
+
+        Return: None
+        """
         self.client.get("/api/registro")
 
     @task(1)
     def login_get(self):
+        """Tarea para solicitar la página de login vía GET
+
+        Return: None
+        """
         self.client.get("/api/login")
 
     @task(1)
     def menu(self):
+        """Tarea para obtener el menú vía GET
+
+        Return: None
+        """
         self.client.get("/api/menu")
 
     @task(3)
     def agregar_producto_carrito(self):
+        """Tarea para agregar un producto aleatorio al carrito vía POST
+
+        Return: None
+        """
         if not self.productos_disponibles:
             print("⚠️ No hay productos disponibles.")
             return
@@ -115,6 +178,10 @@ class MiUsuario(HttpUser):
 
     @task(1)
     def get_productos(self):
+        """Tarea para obtener la lista de productos vía GET
+
+        Return: None
+        """
         response = self.client.get("/api/productos")
         if response.status_code == 200:
             print("✅ Muestreo de productos exitoso")
@@ -123,6 +190,10 @@ class MiUsuario(HttpUser):
 
     @task(2)
     def get_producto_por_id(self):
+        """Tarea para obtener un producto aleatorio por ID vía GET
+
+        Return: None
+        """
         if not self.productos_disponibles:
             return
 
@@ -135,6 +206,10 @@ class MiUsuario(HttpUser):
 
     @task(3)
     def get_un_pedido(self):
+        """Tarea para obtener un pedido aleatorio de la lista vía GET
+
+        Return: None
+        """
         if not self.lista_pedidos:
             print("⚠️ No hay pedidos en la lista para consultar.")
             return
@@ -149,6 +224,10 @@ class MiUsuario(HttpUser):
 
     @task(2)
     def checkout(self):
+        """Tarea para realizar checkout y vaciar carrito
+
+        Return: None
+        """
         with self.client.post("/api/checkout", allow_redirects=False, catch_response=True) as response:
             try:
                 if response.status_code == 200:
@@ -159,11 +238,6 @@ class MiUsuario(HttpUser):
                         print("✅ Pedido creado con ID:", id_pedido)
                         response.success()
                         # Simulación de vaciado del carrito
-                        with self.client.delete("/api/carrito/vaciar", json=data, allow_redirects=False, catch_response=True) as response2:
-                            if self.manejar_redireccion(response2, "✅ Carrito vaciado"):
-                                response2.success()
-                            else:
-                                response2.failure("❌ Fallo al vaciar carrito")
                     elif 'redirect_url' in data:
                         print("⚠️ Sin items de carrito")
                         response.success()
@@ -179,6 +253,10 @@ class MiUsuario(HttpUser):
 
     @task(2)
     def actualizar_producto_put(self):
+        """Tarea para actualizar producto completo vía PUT
+
+        Return: None
+        """
         if not self.productos_disponibles:
             return
 
@@ -201,6 +279,10 @@ class MiUsuario(HttpUser):
 
     @task(2)
     def actualizar_tipo_patch(self):
+        """Tarea para actualizar solo el tipo del producto vía PATCH
+
+        Return: None
+        """
         if not self.productos_disponibles:
             return
 
@@ -218,6 +300,10 @@ class MiUsuario(HttpUser):
 
     @task(1)
     def get_carrito(self):
+        """Tarea para obtener el carrito vía GET
+
+        Return: None
+        """
         with self.client.get("/api/carrito", allow_redirects=False, catch_response=True) as response:
             if self.manejar_redireccion(response, "✅ El carrito se muestra correctamente"):
                 response.success()
@@ -226,13 +312,10 @@ class MiUsuario(HttpUser):
 
     @task(1)
     def delete_carrito(self):
-        """sumary_line
-        
-        Keyword arguments:
-        argument -- description
-        Return: return_description
+        """Tarea para vaciar el carrito vía DELETE
+
+        Return: None
         """
-        
         with self.client.delete("/api/carrito/vaciar", allow_redirects=False, catch_response=True) as response:
             if self.manejar_redireccion(response, "✅ El carrito se vacia correctamente"):
                 response.success()
@@ -241,21 +324,22 @@ class MiUsuario(HttpUser):
 
     @task(1)
     def get_pedidos(self):
-        """sumary_line
-        
-        Keyword arguments:
-        argument -- description
-        Return: return_description
+        """Tarea para obtener la lista de pedidos vía GET
+
+        Return: None
         """
-        
         with self.client.get("/api/pedidos", allow_redirects=False, catch_response=True) as response:
             if self.manejar_redireccion(response, "✅ Los pedidos se muestran correctamente"):
                 response.success()
             else:
                 response.failure("❌ Fallo al obtener pedidos")
-    
+
     @task(1)
     def cerrar_sesion(self):
+        """Tarea para cerrar sesión vía GET
+
+        Return: None
+        """
         with self.client.get("/api/logout", allow_redirects=False, catch_response=True) as response:
             if self.manejar_redireccion(response, "✅ Cierre de sesión correcto"):
                 response.success()
